created the PQLQueryInput component
